Resolved clang format and .hpp file and use pragma
